feat:refactor readOneById test
